refactor: colocate Obsolete handling files
