+++ src/Obsolete/Obsolete.V3.Indicators.cs
