+++ src/Obsolete/Obsolete.V3.Other.cs
