+++ tests/Library/Obsolete/ObsoleteShim.StdDevChannels.Tests.cs
@@ -1,4 +1,5 @@
-namespace Utilities;
+namespace ObsoleteShims;
+#pragma warning disable CS0618
 
 /// <summary>
 /// Tests for obsolete v3 shims whose signatures were narrower than the v2 methods they
@@ -17,49 +18,15 @@ namespace Utilities;
 /// </para>
 /// </remarks>
 [TestClass]
-public class ObsoleteShimNarrowedSignatureTests : TestBase
+public class ObsoleteStdDevChannelsTests : TestBase
 {
-    [TestMethod]
-    public void GetKeltnerAcceptsAFractionalMultiplier()
-    {
-#pragma warning disable CS0618 // exercising the obsolete shim is the point
-        List<KeltnerResult> shim = Bars.GetKeltner(20, 2.5, 10).ToList();
-#pragma warning restore CS0618
-
-        IReadOnlyList<KeltnerResult> expected = Bars.ToKeltner(20, 2.5, 10);
-
-        shim.Should().HaveCount(expected.Count);
-        shim.Select(static r => r.UpperBand).Should().Equal(expected.Select(static r => r.UpperBand));
-        shim.Select(static r => r.Centerline).Should().Equal(expected.Select(static r => r.Centerline));
-        shim.Select(static r => r.LowerBand).Should().Equal(expected.Select(static r => r.LowerBand));
-    }
-
-    [TestMethod]
-    public void GetKeltnerDefaultsMatchToKeltner()
-    {
-        // the bands, not just the centerline: Centerline is the EMA and does not depend
-        // on multiplier at all, so asserting it alone cannot catch a wrong default
-#pragma warning disable CS0618
-        List<KeltnerResult> shim = Bars.GetKeltner().ToList();
-#pragma warning restore CS0618
-
-        IReadOnlyList<KeltnerResult> expected = Bars.ToKeltner();
-
-        shim.Should().HaveCount(expected.Count);
-        shim.Select(static r => r.UpperBand).Should().Equal(expected.Select(static r => r.UpperBand));
-        shim.Select(static r => r.Centerline).Should().Equal(expected.Select(static r => r.Centerline));
-        shim.Select(static r => r.LowerBand).Should().Equal(expected.Select(static r => r.LowerBand));
-    }
-
     [TestMethod]
     public void GetStdDevChannelsTupleAcceptsANullLookback()
     {
         IEnumerable<(DateTime d, double v)> priceTuples
             = Bars.Select(static b => (b.Timestamp, (double)b.Close));
 
-#pragma warning disable CS0618
         List<StdDevChannelsResult> shim = priceTuples.GetStdDevChannels(null, 2).ToList();
-#pragma warning restore CS0618
 
         // null means "the whole series", matching the sibling bar overload
         IReadOnlyList<StdDevChannelsResult> expected
@@ -76,9 +43,7 @@ public class ObsoleteShimNarrowedSignatureTests : TestBase
         IEnumerable<(DateTime d, double v)> priceTuples
             = Bars.Select(static b => (b.Timestamp, (double)b.Close));
 
-#pragma warning disable CS0618
         List<StdDevChannelsResult> shim = priceTuples.GetStdDevChannels().ToList();
-#pragma warning restore CS0618
 
         IReadOnlyList<StdDevChannelsResult> expected
             = Bars.Use(CandlePart.Close).ToStdDevChannels(20, 2);
@@ -91,16 +56,12 @@ public class ObsoleteShimNarrowedSignatureTests : TestBase
     public void GetStdDevChannelsBarOverloadStillAcceptsANullLookback()
     {
         // the overload that was already correct, so the two forms stay in step
-#pragma warning disable CS0618
         List<StdDevChannelsResult> barForm = Bars.GetStdDevChannels(null, 2).ToList();
-#pragma warning restore CS0618
 
         IEnumerable<(DateTime d, double v)> priceTuples
             = Bars.Select(static b => (b.Timestamp, (double)b.Close));
 
-#pragma warning disable CS0618
         List<StdDevChannelsResult> tupleForm = priceTuples.GetStdDevChannels(null, 2).ToList();
-#pragma warning restore CS0618
 
         tupleForm.Select(static r => r.Centerline).Should().Equal(barForm.Select(static r => r.Centerline));
     }
